update some tests
