+++ .gitignore
@@ -20,3 +20,4 @@ src/app/_baked_features.py
 
 __pycache__/
 *.pyc
+.pytest_tmp/
