+++ pytest.ini
@@ -1,5 +1,6 @@
 [pytest]
 testpaths = src/tests
 pythonpath = src
+addopts = --basetemp=.pytest_tmp
 python_files = test_*.py
 python_functions = test_*
