+++ src/tests/conftest.py
@@ -1,28 +1,8 @@
 """Shared pytest fixtures for app-layer filesystem tests."""
 
-from pathlib import Path
-import shutil
-import tempfile
-
 import pytest
 
 from app import edit_session
-
-
-@pytest.fixture
-def tmp_path():
-    """Use a repo-local temp folder when the system temp root is restricted."""
-    parent = Path(__file__).resolve().parents[2] / ".pytest_tmp"
-    parent.mkdir(exist_ok=True)
-    path = Path(tempfile.mkdtemp(prefix="pytest-", dir=parent))
-    try:
-        yield path
-    finally:
-        shutil.rmtree(path, ignore_errors=True)
-        try:
-            parent.rmdir()
-        except OSError:
-            pass
 
 
 @pytest.fixture
