+++ src/tests/test_provenance.py
@@ -1,51 +1,18 @@
 """Tests for source provenance from ini line to UI mesh.
 
-Every `drawindexed` that contributes to a mesh must be traceable back to the
-exact file and line it came from, because an authoring edit has to fan back
-out to all of them. The dedup merge in build_mesh_payload is the dangerous
-part: it collapses several draws into one, and used to keep only the first.
+Every draw that contributes to a mesh must remain traceable to its source file
+and line after parsing and deduplication.
 """
 
-import os, tempfile, struct
+import os
+import tempfile
 
-
-from _corpus import sample_inis, sample_mods
-from core import ini_parser
-from core.ini_parser import parse_sections, merge_sections, build_draw_groups, \
-    extract_resources, extract_toggle_keys, line_source, SrcLine
-from core.mesh_builder import (GeometryBlob, build_mesh_result,
-                               encode_texture_file, split_texture_key,
-                               _deduplicate_draws, _extract_light_mask,
-                               _reconstruct_normal_z)
 from app import mod_loader
-
-
-def geometry_values(blob, reference):
-    start = reference["offset"]
-    end = start + reference["length"]
-    raw = blob.data[start:end]
-    return struct.unpack(f"<{len(raw) // 4}f", raw)
-
-
-def texture_file(key):
-    return split_texture_key(key)[1] if key else key
-
-
-def write(tmp, name, text):
-    p = os.path.join(tmp, name)
-    with open(p, "w", encoding="utf-8") as f:
-        f.write(text)
-    return p
-
-
-def build_mesh_fixture(groups, mod_dir):
-    """Build production-style offset references for a mesh regression."""
-    geometry = GeometryBlob()
-    result = build_mesh_result(groups, mod_dir, geometry=geometry)
-    return result.meshes, geometry
-
-
-# ── SrcLine behaves exactly like str ─────────────────────────────────────────
+from core.ini_parser import (SrcLine, build_draw_groups, extract_resources,
+                             extract_toggle_keys, line_source, merge_sections,
+                             parse_sections)
+from core.mesh_builder import _deduplicate_draws
+from _provenance_support import write
 
 def test_srcline_is_a_str():
     s = SrcLine("drawindexed = 1, 2, 3", "C:\\a.ini", 42, "TextureOverrideBody")
@@ -57,8 +24,6 @@ def test_srcline_is_a_str():
                              "section": "TextureOverrideBody"}), ("line_source returns file/line/section")
     assert (line_source("plain string") is None), ("line_source of a plain str is None")
 
-
-# ── parse_sections records the real line numbers ─────────────────────────────
 
 INI = """; leading comment
 
@@ -146,8 +111,6 @@ def test_toggle_key_provenance():
         assert (info and info["section"] == "KeySwap"), ("toggle key knows its section")
 
 
-# ── the dedup merge must not lose contributing lines ─────────────────────────
-
 SHARED_INI = """[Constants]
 global persist $swapvar = 0
 
@@ -227,9 +190,6 @@ def test_deduplicate_preserves_buffer_identity():
     assert (len(merged) == 2), (f"different buffer bindings do not collapse by range alone (got {len(merged)})")
 
 
-# ── cross-ini label collision (real bug: BellyDancer_mod + HairPin_mod both
-# defining their own unrelated [TextureOverrideComponent0]) ─────────────────
-
 COMPONENT0_INI = """[CommandListShared]
 ib = ResourceBodyIB
 vb0 = ResourcePos
@@ -275,1302 +235,3 @@ def test_cross_ini_component_collision_recovered():
         components = {e.get("component") for e in mesh_entries.values()}
         assert (components == {"Component0"}), (f"both display as the same clean name -- no '_2' suffix leaks into the UI "
               f"(got {components}); the per-source header is what actually disambiguates them")
-
-
-
-# ── end-to-end against real mods ─────────────────────────────────────────────
-
-def _corpus_case_real_mods():
-    mods = sample_mods(15, seed=11)
-    if not mods:
-        print("SKIP  no local mod libraries found")
-        return
-
-    total_meshes = missing = multi = 0
-    bad_line = bad_file = 0
-    skipped = 0
-
-    for mod in mods:
-        payload = mod_loader.load_mod(mod)
-        if "error" in payload:
-            # "Master" dispatch inis in AllInOne folders hold only menus and
-            # namespace switches, never geometry — nothing to trace.
-            if "No mesh geometry" in payload["error"]:
-                skipped += 1
-            else:
-                assert (False), (f"{os.path.basename(mod)} failed to load: "
-                             f"{payload['error'].strip().splitlines()[-1][:100]}")
-            continue
-        for name, entry in payload.get("meshes", {}).items():
-            if not isinstance(entry, dict):
-                continue
-            total_meshes += 1
-            srcs = entry.get("sources") or []
-            if not srcs:
-                missing += 1
-                continue
-            if len(srcs) > 1:
-                multi += 1
-            for s in srcs:
-                ini_abs = os.path.join(mod, s["ini"] or "")
-                if not os.path.isfile(ini_abs):
-                    bad_file += 1
-                    continue
-                with open(ini_abs, encoding="utf-8", errors="ignore") as f:
-                    all_lines = f.readlines()
-                n = s["line"]
-                if not (n and 1 <= n <= len(all_lines)):
-                    bad_line += 1
-                    continue
-                text = all_lines[n - 1].strip().lower()
-                # Normally the exact drawindexed. A section with no drawindexed
-                # at all is drawn by reading the whole IB, and is anchored to
-                # the first significant line of its section instead.
-                if not (text.startswith("drawindexed")
-                        or _anchors_section(all_lines, n, s["section"])):
-                    bad_line += 1
-
-    print(f"      {len(mods)} mods ({skipped} geometry-free), {total_meshes} meshes, "
-          f"{multi} multi-source, {missing} without provenance")
-    assert (total_meshes > 0), ("real mods produced meshes")
-    assert (missing == 0), (f"every mesh has provenance (missing={missing})")
-    assert (bad_file == 0), (f"every recorded ini path resolves (bad={bad_file})")
-    assert (bad_line == 0), (f"every recorded line is a drawindexed or a section anchor (bad={bad_line})")
-    assert (multi > 0), (f"at least one mesh has several sources (got {multi})")
-
-
-def _corpus_case_diffuse_resolution_corpus_sweep():
-    """The execution-order diffuse fix (per-draw `texture_default_file` /
-    `diffuse_pool_files`, replacing one static tex_key per component) must
-    not crash or misbehave across the real corpus, and must actually change
-    resolution for the multi-diffuse sections it exists to fix. For a section
-    referencing at most one diffuse, that diffuse must still apply to AT
-    LEAST ONE draw (a draw legitimately preceding the assignment in file
-    order gets None instead, which is fine) -- if it resolves but zero draws
-    ever get it, that's the whole point of resolving lost.
-    """
-    inis = sample_inis(400, seed=17)
-    if not inis:
-        print("SKIP  no local mod libraries found")
-        return
-
-    sections = draws = multi_diffuse_sections = differing_draws = 0
-    unresolved_pool_entries = crashes = 0
-    single_diffuse_mismatches = 0
-
-    for ini_path in inis:
-        try:
-            secs = merge_sections([ini_path])
-            groups = build_draw_groups(secs, extract_resources(secs))
-        except Exception as e:
-            crashes += 1
-            assert (False), (f"{ini_path}: build_draw_groups crashed: {e}")
-            continue
-
-        for grp in groups:
-            sections += 1
-            pool = grp.get("diffuse_pool_files") or []
-            for entry in pool:
-                if not entry.get("file"):
-                    unresolved_pool_entries += 1
-            defaults = set()
-            for d in grp["draws"]:
-                draws += 1
-                defaults.add(d.get("texture_default_file"))
-            if len(pool) >= 2:
-                multi_diffuse_sections += 1
-                if len(defaults) > 1:
-                    differing_draws += 1
-            elif grp.get("diffuse_file"):
-                # At most one diffuse ever resolves. A draw legitimately
-                # precedes the (single) assignment in file order and gets
-                # None for it -- real corpus examples: Caesar/Promeia/Aria
-                # sections with several drawindexed lines before their
-                # Resource\...\Diffuse line. So not every draw need equal
-                # the old static value -- but AT LEAST ONE must, or the
-                # diffuse never actually applies to anything despite
-                # resolving. Real regression this catches: a section with NO
-                # drawindexed line at all (implicit whole-buffer draw) whose
-                # synthetic placeholder used to hardcode an empty variant
-                # list, resolving to None even though e.g. a bare `ps-t0 =`
-                # names a real diffuse (Beidou.ini's
-                # TextureOverrideBeidouBody) -- there `defaults` is just
-                # {None} because there's only the one placeholder draw.
-                old_static = grp["diffuse_file"]
-                if old_static not in defaults:
-                    single_diffuse_mismatches += 1
-
-    print(f"      {len(inis)} inis, {sections} sections, {draws} draws, "
-          f"{multi_diffuse_sections} multi-diffuse sections "
-          f"({differing_draws} with draws that actually resolve differently), "
-          f"{unresolved_pool_entries} unresolved pool entries")
-    assert (sections > 0), ("the corpus sample produced draw groups")
-    assert (crashes == 0), (f"build_draw_groups never raises on real inis (got {crashes})")
-    assert (unresolved_pool_entries == 0), (f"every diffuse_pool_files entry resolves to a filename "
-          f"(got {unresolved_pool_entries} unresolved)")
-    assert (single_diffuse_mismatches == 0), (f"a section referencing at most one diffuse resolves identically "
-          f"to the old static-per-component model (got "
-          f"{single_diffuse_mismatches} mismatches)")
-
-
-def _anchors_section(all_lines, n, section):
-    """True if line n is the first significant line of `section`."""
-    header = f"[{(section or '').lower()}]"
-    i = n - 2
-    while i >= 0:
-        t = all_lines[i].strip().lower()
-        if t.startswith("[") and t.endswith("]"):
-            return t == header
-        if t and not t.startswith(";"):
-            return False
-        i -= 1
-    return False
-
-
-# ── a mod ini may reach a sibling assets folder, but not roam the disk ───────
-
-def _traversal_mod(tmp, pos_filename):
-    """A minimal one-draw mod whose vb0 filename is `pos_filename`."""
-    ini_text = f"""[Constants]
-global persist $swapvar = 0
-
-[TextureOverrideBodyBlend]
-ib = ResourceBodyIB
-vb0 = ResourcePos
-vb1 = ResourceTc
-drawindexed = 100, 0, 0
-
-[ResourceBodyIB]
-filename = body.ib
-format = DXGI_FORMAT_R32_UINT
-
-[ResourcePos]
-filename = {pos_filename}
-stride = 40
-
-[ResourceTc]
-filename = tc.buf
-stride = 20
-"""
-    path = write(tmp, "mod.ini", ini_text)
-    for buf in ("body.ib", "tc.buf"):
-        open(os.path.join(tmp, buf), "wb").write(b"\0" * 4096)
-    secs = merge_sections([path])
-    groups = build_draw_groups(secs, extract_resources(secs))
-    meshes, _geometry = build_mesh_fixture(groups, tmp)
-    return meshes
-
-
-def _resource_case_resource_path_may_reach_a_sibling_folder():
-    """`filename = ..\\resources\\x.buf` is how mods share assets between the
-    ini's folder and its neighbours -- it has to resolve."""
-    with tempfile.TemporaryDirectory() as tmp:
-        mod = os.path.join(tmp, "mod")
-        shared = os.path.join(tmp, "shared")
-        os.makedirs(mod); os.makedirs(shared)
-        open(os.path.join(shared, "pos.buf"), "wb").write(b"\1" * 4096)
-
-        meshes = _traversal_mod(mod, "../shared/pos.buf")
-        assert (len(meshes) == 1), (f"a resource one folder above the ini is read (got {list(meshes)})")
-
-
-def _resource_case_absolute_resource_path_blocked():
-    """The mod folder is untrusted, downloaded content: a crafted `filename`
-    naming an absolute path must not be read."""
-    with tempfile.TemporaryDirectory() as outside, tempfile.TemporaryDirectory() as tmp:
-        secret = os.path.join(outside, "secret.buf")
-        with open(secret, "wb") as f:
-            f.write(b"\1" * 4096)
-
-        meshes = _traversal_mod(tmp, secret.replace(os.sep, "/"))
-        assert (not meshes), (f"absolute resource path is refused (got {list(meshes)})")
-
-
-def _resource_case_deep_resource_path_traversal_blocked():
-    """`..` is allowed, but only a few levels up -- not far enough to walk out
-    of the mod library and into the user's own files."""
-    with tempfile.TemporaryDirectory() as tmp:
-        secret = os.path.join(tmp, "secret.buf")
-        with open(secret, "wb") as f:
-            f.write(b"\1" * 4096)
-        mod = os.path.join(tmp, "a", "b", "c", "d", "e")
-        os.makedirs(mod)
-
-        meshes = _traversal_mod(mod, "../../../../../secret.buf")
-        assert (not meshes), (f"a resource far above the mod folder is refused (got {list(meshes)})")
-
-
-def _resource_case_root_texture_picker_accepts_windows_case_variation():
-    """The two native dialogs may spell the same Windows folder differently."""
-    if os.name != "nt":
-        return
-    from PIL import Image
-
-    with tempfile.TemporaryDirectory(prefix="TexturePickerCase") as tmp:
-        path = os.path.join(tmp, "RootDiffuse.png")
-        Image.new("RGB", (1, 1), (255, 0, 0)).save(path)
-
-        # Root-level files retain a role-aware key; the differently-cased path
-        # must still pass containment validation on case-insensitive Windows.
-        result = encode_texture_file(tmp.swapcase(), path)
-        assert (not result.get("error")), (f"root-level picked texture accepts equivalent path casing ({result})")
-        assert (result.get("tex_key") == "diffuse::RootDiffuse.png"
-              and result.get("file") == "RootDiffuse.png"), (f"root-level picked texture keeps role and source path ({result})")
-
-
-def _resource_case_toggle_panel_provenance():
-    mods = sample_mods(15, seed=11)
-    if not mods:
-        return
-    checked = bad = 0
-    for mod in mods:
-        payload = mod_loader.load_mod(mod)
-        for section, info in (payload.get("controls", {}).get("toggles") or {}).items():
-            checked += 1
-            ini = info.get("ini")
-            if not ini or not os.path.isfile(os.path.join(mod, ini)):
-                bad += 1
-                continue
-            secs = parse_sections(os.path.join(mod, ini))
-            if info.get("section") not in secs:
-                bad += 1
-    print(f"      {checked} toggle sections checked")
-    assert (checked > 0), ("real mods produced toggle sections")
-    assert (bad == 0), (f"every toggle resolves to a real section in a real file (bad={bad})")
-
-
-# ── mid-section `ib =` reassignment must not be lost
-
-IB_REASSIGN_INI = """[TextureOverrideBodyBlend]
-ib = ResourceBodyHeadIB
-vb0 = ResourcePos
-vb1 = ResourceTc
-drawindexed = 100, 0, 0
-ib = ResourceBodyDressIB
-drawindexed = 100, 0, 0
-
-[ResourceBodyHeadIB]
-filename = head.ib
-format = DXGI_FORMAT_R32_UINT
-
-[ResourceBodyDressIB]
-filename = dress.ib
-format = DXGI_FORMAT_R32_UINT
-
-[ResourcePos]
-filename = pos.buf
-stride = 40
-
-[ResourceTc]
-filename = tc.buf
-stride = 20
-"""
-
-
-def _mesh_case_mid_section_ib_reassignment_ini_parser():
-    with tempfile.TemporaryDirectory() as tmp:
-        path = write(tmp, "mod.ini", IB_REASSIGN_INI)
-        secs = merge_sections([path])
-        groups = build_draw_groups(secs, extract_resources(secs))
-        assert (len(groups) == 1), (f"one draw group built (got {len(groups)})")
-        draws = groups[0]["draws"]
-        assert (len(draws) == 2), (f"both drawindexed lines kept (got {len(draws)})")
-        assert (groups[0]["ib_file"] == "head.ib"), (f"group's default ib is the section's first-seen one (got {groups[0]['ib_file']})")
-        assert (draws[0].get("ib_file") is None), ("first draw has no override -- reads the group's default ib (head.ib)")
-        assert (draws[1].get("ib_file") == "dress.ib"), (f"second draw carries the reassigned ib (got {draws[1].get('ib_file')})")
-
-
-def _mesh_case_mid_section_ib_reassignment_mesh_builder():
-    """End-to-end: build_mesh_result must read each draw's indices from its
-    own reassigned ib file, and must not merge two draws that happen to share
-    (start, count) but actually come from different index buffers."""
-    with tempfile.TemporaryDirectory() as tmp:
-        path = write(tmp, "mod.ini", IB_REASSIGN_INI)
-        open(os.path.join(tmp, "head.ib"), "wb").write(struct.pack("<3I", 10, 11, 12))
-        open(os.path.join(tmp, "dress.ib"), "wb").write(struct.pack("<3I", 20, 21, 22))
-        # 32 unique, identifiable vertices: vertex i sits at position (i, i, i)
-        with open(os.path.join(tmp, "pos.buf"), "wb") as f:
-            for i in range(32):
-                f.write(struct.pack("<3f", float(i), float(i), float(i)) + b"\0" * 28)
-        open(os.path.join(tmp, "tc.buf"), "wb").write(b"\0" * 20 * 32)
-
-        secs = merge_sections([path])
-        groups = build_draw_groups(secs, extract_resources(secs))
-        # both draws are (start=0, count=100) in the ini above -- shrink to the
-        # 3 real indices we actually wrote so read_indices has something to read
-        for d in groups[0]["draws"]:
-            d["count"] = 3
-        meshes, geometry = build_mesh_fixture(groups, tmp)
-        assert (len(meshes) == 2), (f"both draws survive as distinct meshes, not merged "
-                                 f"(got {len(meshes)})")
-
-        def _verts(entry):
-            pos = geometry_values(geometry, entry["pos"])
-            return sorted(round(pos[i]) for i in range(0, len(pos), 3))
-
-        vert_sets = sorted(_verts(e) for e in meshes.values())
-        assert (vert_sets == [[10, 11, 12], [20, 21, 22]]), (f"each mesh's own vertices come from its own reassigned ib (got {vert_sets})")
-
-
-# ── a mid-section `ib =` reassignment paired with `vb0/vb1 =` (a "cross IB/VB"
-#    swap some GIMI mods use to draw a wholly different mesh's buffers for a
-#    handful of draws) must resolve the NEW mesh's own position/texcoord too,
-#    not just its own ib -- otherwise the new ib's indices get read against
-#    the old mesh's (unrelated, often shorter) position buffer, producing
-#    garbage vertices. The literal `vb0=`/`vb1=` values paired with the
-#    reassignment are deliberately NOT the source of truth here (real mods
-#    sometimes set them to a `= copy vb0` runtime GPU snapshot with no
-#    filename at all) -- the fix re-derives the buffers from the reassigned
-#    `ib`'s own component, same as a normal group's defaults are resolved.
-
-CROSS_IB_VB_INI = """[TextureOverrideSBSBlend]
-vb0 = ResourceSBSPosition
-vb1 = ResourceSBSTexcoord
-
-[TextureOverrideXBSBlend]
-vb0 = ResourceXBSPosition
-vb1 = ResourceXBSTexcoord
-
-[TextureOverrideSBSA]
-ib = ResourceSBSAIB
-drawindexed = 3, 0, 0
-ib = ResourceXBSAIB
-vb0 = ResourceXBSCrossIBVB
-vb1 = ResourceXBSTexcoord
-drawindexed = 3, 0, 0
-
-[ResourceXBSCrossIBVB]
-
-[ResourceSBSAIB]
-filename = sbsA.ib
-format = DXGI_FORMAT_R32_UINT
-
-[ResourceXBSAIB]
-filename = xbsA.ib
-format = DXGI_FORMAT_R32_UINT
-
-[ResourceSBSPosition]
-filename = sbsPos.buf
-stride = 40
-
-[ResourceSBSTexcoord]
-filename = sbsTc.buf
-stride = 20
-
-[ResourceXBSPosition]
-filename = xbsPos.buf
-stride = 40
-
-[ResourceXBSTexcoord]
-filename = xbsTc.buf
-stride = 20
-"""
-
-
-def _mesh_case_cross_ib_vb_reassignment_ini_parser():
-    with tempfile.TemporaryDirectory() as tmp:
-        path = write(tmp, "mod.ini", CROSS_IB_VB_INI)
-        secs = merge_sections([path])
-        groups = build_draw_groups(secs, extract_resources(secs))
-        assert (len(groups) == 1), (f"one draw group built (got {len(groups)})")
-        draws = groups[0]["draws"]
-        assert (len(draws) == 2), (f"both drawindexed lines kept (got {len(draws)})")
-        assert ("position_file" not in draws[0] and "texcoord_file" not in draws[0]), ("first draw has no override -- reads the group's default SBS buffers")
-        assert (draws[1].get("position_file") == "xbsPos.buf" and
-              draws[1].get("texcoord_file") == "xbsTc.buf"), (f"second draw carries its own reassigned vertex buffers "
-              f"(got {draws[1].get('position_file')}, {draws[1].get('texcoord_file')})")
-
-
-def _mesh_case_cross_ib_vb_reassignment_mesh_builder():
-    """End-to-end: the second draw's indices must decode against its own
-    reassigned XBS position buffer, not the group's default (shorter) SBS one
-    -- reading against the wrong buffer either raises IndexError or silently
-    collapses out-of-range vertices to the origin, which is exactly the
-    "triangles fly everywhere" corruption this guards against."""
-    with tempfile.TemporaryDirectory() as tmp:
-        path = write(tmp, "mod.ini", CROSS_IB_VB_INI)
-        open(os.path.join(tmp, "sbsA.ib"), "wb").write(struct.pack("<3I", 0, 1, 2))
-        open(os.path.join(tmp, "xbsA.ib"), "wb").write(struct.pack("<3I", 5, 6, 7))
-        # SBS has only 4 vertices; XBS has 8, at a different scale, so decoding
-        # against the wrong one is either out-of-range or visibly wrong.
-        with open(os.path.join(tmp, "sbsPos.buf"), "wb") as f:
-            for i in range(4):
-                f.write(struct.pack("<3f", float(i), float(i), float(i)) + b"\0" * 28)
-        with open(os.path.join(tmp, "xbsPos.buf"), "wb") as f:
-            for i in range(8):
-                v = float(i * 10)
-                f.write(struct.pack("<3f", v, v, v) + b"\0" * 28)
-        open(os.path.join(tmp, "sbsTc.buf"), "wb").write(b"\0" * 20 * 4)
-        open(os.path.join(tmp, "xbsTc.buf"), "wb").write(b"\0" * 20 * 8)
-
-        secs = merge_sections([path])
-        groups = build_draw_groups(secs, extract_resources(secs))
-        meshes, geometry = build_mesh_fixture(groups, tmp)
-        assert (len(meshes) == 2), (f"both draws survive as distinct meshes (got {len(meshes)})")
-
-        def _verts(entry):
-            pos = geometry_values(geometry, entry["pos"])
-            return sorted(round(pos[i]) for i in range(0, len(pos), 3))
-
-        vert_sets = sorted(_verts(e) for e in meshes.values())
-        assert (vert_sets == [[0, 1, 2], [50, 60, 70]]), (f"the reassigned draw decodes against its own XBS position buffer, "
-              f"not a collapsed/garbage read of the SBS one (got {vert_sets})")
-
-
-# ── A ZZMI shape-key output can be an empty writable Resource whose file-backed
-#    rest pose is identified only by `ResourceX = copy ResourceXBase`.
-
-RUNTIME_POSITION_COPY_INI = """[TextureOverrideBodyBlend]
-vb0 = ResourceBodyPosition
-
-[TextureOverrideBodyTexcoord]
-vb1 = ResourceBodyTexcoord
-
-[TextureOverrideLegsBlend]
-vb0 = ResourceLegsPosition
-
-[TextureOverrideLegsTexcoord]
-vb1 = ResourceLegsTexcoord
-
-[TextureOverrideLegsA]
-ib = ResourceLegsAIB
-drawindexed = 3, 0, 0
-ib = ResourceBodyAIB
-vb0 = ResourceBodyRuntimeSnapshot
-vb1 = ResourceBodyTexcoord
-drawindexed = 3, 0, 0
-
-[Present]
-ResourceBodyPosition = copy ResourceBodyPositionBase
-ResourceLegsPosition = copy ResourceLegsPositionBase
-
-[ResourceBodyPosition]
-[ResourceLegsPosition]
-[ResourceBodyRuntimeSnapshot]
-
-[ResourceBodyPositionBase]
-filename = bodyBase.buf
-stride = 40
-
-[ResourceLegsPositionBase]
-filename = legsBase.buf
-stride = 40
-
-[ResourceBodyTexcoord]
-filename = bodyTc.buf
-stride = 20
-
-[ResourceLegsTexcoord]
-filename = legsTc.buf
-stride = 20
-
-[ResourceBodyAIB]
-filename = bodyA.ib
-format = DXGI_FORMAT_R32_UINT
-
-[ResourceLegsAIB]
-filename = legsA.ib
-format = DXGI_FORMAT_R32_UINT
-"""
-
-
-def _resource_case_runtime_position_copy_resolution():
-    with tempfile.TemporaryDirectory() as tmp:
-        path = write(tmp, "mod.ini", RUNTIME_POSITION_COPY_INI)
-        secs = merge_sections([path])
-        groups = build_draw_groups(secs, extract_resources(secs))
-
-        assert (len(groups) == 1), (f"runtime position resources no longer drop the "
-                                f"draw group (got {len(groups)})")
-        if not groups:
-            return
-        group = groups[0]
-        assert (group["position_file"] == "legsBase.buf"), (f"group position follows the explicit Legs -> LegsBase copy "
-              f"(got {group['position_file']})")
-        assert (group["draws"][1].get("position_file") == "bodyBase.buf"), (f"reassigned Body draw follows the explicit Body -> BodyBase copy "
-              f"(got {group['draws'][1].get('position_file')})")
-
-
-LL_SKELETON_OUTPUT_INI = """[TextureOverrideBodyBlend]
-vb2 = ResourceBodyBlend
-
-[TextureOverrideBodyTexcoord]
-vb1 = ResourceBodyTexcoord
-
-[TextureOverrideBodyA]
-run = CommandListBodyA
-
-[CommandListBodyA]
-ib = ResourceBodyAIB
-run = CommandListLLSkeletonSkin_Body
-drawindexed = 3, 0, 0
-
-[CommandListLLSkeletonSkin_Body]
-cs-t1 = ref ResourceBodyPosition
-cs-t2 = ref ResourceBodyBlend
-cs-u0 = ref ResourceLLSkelOutput_Body
-cs-u0 = null
-cs-t1 = null
-cs-t2 = null
-vb0 = ref ResourceLLSkelOutput_Body
-
-[ResourceBodyPosition]
-filename = bodyPosition.buf
-stride = 40
-
-[ResourceBodyBlend]
-filename = bodyBlend.buf
-stride = 32
-
-[ResourceBodyTexcoord]
-filename = bodyTexcoord.buf
-stride = 24
-
-[ResourceLLSkelOutput_Body]
-
-[ResourceBodyAIB]
-filename = bodyA.ib
-format = DXGI_FORMAT_R32_UINT
-"""
-
-
-def _resource_case_ll_skeleton_compute_output_uses_rest_position():
-    with tempfile.TemporaryDirectory() as tmp:
-        path = write(tmp, "mod.ini", LL_SKELETON_OUTPUT_INI)
-        secs = merge_sections([path])
-        groups = build_draw_groups(secs, extract_resources(secs))
-
-        assert (len(groups) == 1), (f"LL compute-skinned body is retained (got {len(groups)})")
-        if not groups:
-            return
-        group = groups[0]
-        assert (group["position_file"] == "bodyPosition.buf"), (f"runtime LL output resolves to its cs-t1 rest position "
-              f"(got {group['position_file']})")
-        assert (group["texcoord_file"] == "bodyTexcoord.buf"), (f"runtime LL output keeps the sibling texcoord binding "
-              f"(got {group['texcoord_file']})")
-
-
-AUXILIARY_MAPS_INI = """[Constants]
-global $detail = 0
-global $metal = 0
-
-[KeyDetail]
-type = cycle
-$detail = 0,1
-
-[KeyMetal]
-type = cycle
-$metal = 0,1
-
-[TextureOverrideBodyBlend]
-vb0 = ResourceBodyPosition
-vb1 = ResourceBodyTexcoord
-
-[TextureOverrideBodyA]
-ib = ResourceBodyAIB
-if $detail == 0
-Resource\\ZZMI\\NormalMap = ref ResourceNormalA
-else
-Resource\\ZZMI\\NormalMap = ref ResourceNormalB
-endif
-Resource\\ZZMI\\LightMap = ref ResourceLight
-if $metal == 1
-Resource\\ZZMI\\MaterialMap = ref ResourceMaterial
-endif
-drawindexed = 3, 0, 0
-
-[ResourceBodyPosition]
-filename = pos.buf
-stride = 40
-[ResourceBodyTexcoord]
-filename = tc.buf
-stride = 20
-[ResourceBodyAIB]
-filename = body.ib
-format = DXGI_FORMAT_R32_UINT
-[ResourceNormalA]
-filename = normal-a.dds
-[ResourceNormalB]
-filename = normal-b.dds
-[ResourceLight]
-filename = light.dds
-[ResourceMaterial]
-filename = material.dds
-"""
-
-
-def _material_case_authored_auxiliary_material_maps():
-    with tempfile.TemporaryDirectory() as tmp:
-        path = write(tmp, "mod.ini", AUXILIARY_MAPS_INI)
-        secs = merge_sections([path])
-        groups = build_draw_groups(secs, extract_resources(secs))
-
-        assert (len(groups) == 1), (f"auxiliary-map fixture builds (got {len(groups)})")
-        if not groups:
-            return
-        draw = groups[0]["draws"][0]
-        normals = draw.get("normal_map_variants") or []
-        assert ([v["file"] for v in normals] == ["normal-a.dds", "normal-b.dds"]), (f"conditional normal maps retain both authored branches (got {normals})")
-        assert (draw.get("light_map_default_file") == "light.dds"), (f"unconditional light map becomes the draw default "
-              f"(got {draw.get('light_map_default_file')})")
-        materials = draw.get("material_map_variants") or []
-        assert (len(materials) == 1 and materials[0]["file"] == "material.dds"
-              and materials[0]["conditions"]), (f"a conditional-only material map retains a no-map fallback "
-              f"(got {materials})")
-
-
-def _material_case_direct_ps_t_auxiliary_material_maps():
-    direct = AUXILIARY_MAPS_INI
-    for old, new in (
-            ("ResourceNormalA", "ResourceBodyNormalMapA"),
-            ("ResourceNormalB", "ResourceBodyNormalMapB"),
-            ("ResourceLight", "ResourceBodyLightMap"),
-            ("ResourceMaterial", "ResourceBodyMaterialMap")):
-        direct = direct.replace(old, new)
-    direct = direct.replace(
-        r"Resource\ZZMI\NormalMap = ref ", "ps-t1 = ")
-    direct = direct.replace(
-        r"Resource\ZZMI\LightMap = ref ", "ps-t2 = ")
-    direct = direct.replace(
-        r"Resource\ZZMI\MaterialMap = ref ", "ps-t3 = ")
-
-    with tempfile.TemporaryDirectory() as tmp:
-        path = write(tmp, "mod.ini", direct)
-        secs = merge_sections([path])
-        groups = build_draw_groups(secs, extract_resources(secs))
-
-        assert (len(groups) == 1), (f"direct ps-t auxiliary fixture builds (got {len(groups)})")
-        if not groups:
-            return
-        draw = groups[0]["draws"][0]
-        normals = draw.get("normal_map_variants") or []
-        assert ([v["file"] for v in normals] ==
-              ["normal-a.dds", "normal-b.dds"]), (f"direct ps-t normal maps retain both branches (got {normals})")
-        assert (draw.get("light_map_default_file") == "light.dds"), (f"direct ps-t light map becomes the draw default "
-              f"(got {draw.get('light_map_default_file')})")
-        materials = draw.get("material_map_variants") or []
-        assert (len(materials) == 1 and materials[0]["file"] == "material.dds"
-              and materials[0]["conditions"]), (f"direct ps-t material map retains its condition (got {materials})")
-
-
-def _material_case_two_channel_normal_reconstructs_z():
-    from PIL import Image
-    source = Image.new("RGB", (2, 1))
-    source.putdata([(128, 128, 0), (255, 128, 0)])
-    rebuilt = _reconstruct_normal_z(source)
-    pixels = [rebuilt.getpixel((x, 0)) for x in range(2)]
-    assert (pixels[0][2] == 255), (f"a flat XY normal reconstructs a forward-facing Z (got {pixels[0]})")
-    assert (127 <= pixels[1][2] <= 129), (f"a full-strength X normal reconstructs a near-zero Z (got {pixels[1]})")
-
-
-def _material_case_packed_light_map_uses_blue_mask_without_colour_cast():
-    from PIL import Image
-    source = Image.new("RGB", (1, 1), (210, 12, 94))
-    assert (_extract_light_mask(source).getpixel((0, 0)) == (94, 94, 94)), ("packed LightMap blue is exposed as a neutral scalar mask")
-
-
-# ── `handling = skip` with no `drawindexed` line at all means "suppress the
-#    original draw and replace it with nothing", NOT "draw the whole ib".
-#    Only a section that omits `handling = skip` gets the implicit
-#    whole-buffer-draw fallback.
-
-HANDLING_SKIP_INI = """[TextureOverrideBodyBlend]
-vb0 = ResourcePos
-vb1 = ResourceTc
-
-[TextureOverrideBodyA]
-ib = ResourceBodyAIB
-drawindexed = 100, 0, 0
-
-[TextureOverrideBodyB]
-handling = skip
-ib = ResourceBodyBIB
-
-[TextureOverrideBodyC]
-ib = ResourceBodyCIB
-
-[ResourceBodyAIB]
-filename = bodyA.ib
-format = DXGI_FORMAT_R32_UINT
-
-[ResourceBodyBIB]
-filename = bodyB.ib
-format = DXGI_FORMAT_R32_UINT
-
-[ResourceBodyCIB]
-filename = bodyC.ib
-format = DXGI_FORMAT_R32_UINT
-
-[ResourcePos]
-filename = pos.buf
-stride = 40
-
-[ResourceTc]
-filename = tc.buf
-stride = 20
-"""
-
-
-def _mesh_case_handling_skip_with_no_drawindexed_draws_nothing():
-    with tempfile.TemporaryDirectory() as tmp:
-        path = write(tmp, "mod.ini", HANDLING_SKIP_INI)
-        secs = merge_sections([path])
-        groups = build_draw_groups(secs, extract_resources(secs))
-        names = {g["display_name"] for g in groups}
-        assert ("BodyA" in names), ("the section with an explicit drawindexed still draws")
-        assert ("BodyB" not in names), (f"a handling=skip section with NO drawindexed draws nothing at all "
-              f"(got groups: {sorted(names)})")
-        assert ("BodyC" in names), ("a section with no handling=skip still gets the implicit whole-ib draw")
-
-
-# ── a draw section's `ib=` can name a component that's a per-part suffix
-#    (e.g. "...Head"/"...Body") of the shared Position/Blend/Texcoord
-#    component -- and the shared component's own name can itself end in an
-#    uppercase abbreviation (e.g. Genshin's "CN" outfit-variant suffix), which
-#    broke the CamelCase-word-strip fallback's assumption that a lowercase
-#    letter always precedes the final word.
-
-COMPONENT_ABBREV_SUFFIX_INI = """[TextureOverrideXCNPosition]
-vb0 = ResourceXCNPosition
-
-[TextureOverrideXCNBlend]
-vb1 = ResourceXCNBlend
-
-[TextureOverrideXCNTexcoord]
-vb1 = ResourceXCNTexcoord
-
-[TextureOverrideXCNHead]
-ib = ResourceXCNHeadIB
-drawindexed = 100, 0, 0
-
-[ResourceXCNPosition]
-filename = pos.buf
-stride = 40
-
-[ResourceXCNBlend]
-filename = blend.buf
-stride = 32
-
-[ResourceXCNTexcoord]
-filename = tc.buf
-stride = 12
-
-[ResourceXCNHeadIB]
-filename = head.ib
-format = DXGI_FORMAT_R32_UINT
-"""
-
-
-def _mesh_case_component_name_ending_in_uppercase_abbreviation():
-    with tempfile.TemporaryDirectory() as tmp:
-        path = write(tmp, "mod.ini", COMPONENT_ABBREV_SUFFIX_INI)
-        secs = merge_sections([path])
-        groups = build_draw_groups(secs, extract_resources(secs))
-        names = {g["display_name"]: g for g in groups}
-        assert ("XCNHead" in names), (f"the Head draw resolves its buffers via the shared XCN component "
-              f"(got groups: {sorted(names)})")
-        if "XCNHead" in names:
-            g = names["XCNHead"]
-            assert (g["position_file"] == "pos.buf" and g["texcoord_file"] == "tc.buf"), (f"resolved to the shared component's own buffers "
-                  f"(got {g['position_file']}, {g['texcoord_file']})")
-
-
-RUN_CHAIN_INI = """[Constants]
-global persist $naked = 0
-global persist $flag = 0
-
-[KeyNaked]
-key = n
-type = cycle
-$naked = 0,1
-
-[KeyFlag]
-key = f
-type = cycle
-$flag = 0,1
-
-[TextureOverrideBodyBlend]
-ib = ResourceBodyIB
-vb0 = ResourcePos
-vb1 = ResourceTc
-if $naked == 0
-drawindexed = 100, 0, 0
-run = CustomShaderOuter
-endif
-
-[CustomShaderOuter]
-run = CommandListTransparent
-
-[CommandListTransparent]
-if $flag == 0
-drawindexed = 50, 200, 0
-endif
-
-[ResourceBodyIB]
-filename = body.ib
-format = DXGI_FORMAT_R32_UINT
-
-[ResourcePos]
-filename = pos.buf
-stride = 40
-
-[ResourceTc]
-filename = tc.buf
-stride = 20
-"""
-
-
-def _visible(conds, bindings):
-    if conds == []:
-        return True
-    return any(all((bindings.get(c["var"]) == c["value"]) != c["negate"] for c in group)
-               for group in conds)
-
-
-def _resource_case_run_inlines_nested_commandlist_draws():
-    with tempfile.TemporaryDirectory() as tmp:
-        path = write(tmp, "mod.ini", RUN_CHAIN_INI)
-        secs = merge_sections([path])
-        groups = build_draw_groups(secs, extract_resources(secs))
-        assert (len(groups) == 1), (f"one draw group built (got {len(groups)})")
-        draws = groups[0]["draws"]
-        assert (len(draws) == 2), (f"the run=-chained drawindexed is inlined alongside the direct one "
-              f"(got {len(draws)})")
-        by_count = {d["count"]: d for d in draws}
-        assert (100 in by_count and 50 in by_count), (f"both the direct and run=-chained draws are present (got {sorted(by_count)})")
-
-        chained = by_count[50]
-        assert (_visible(chained["conditions"], {"naked": "0", "flag": "0"})), ("chained draw visible when both naked==0 and flag==0")
-        assert (not _visible(chained["conditions"], {"naked": "1", "flag": "0"})), ("chained draw hidden when the caller's own gate (naked==0) fails")
-        assert (not _visible(chained["conditions"], {"naked": "0", "flag": "1"})), ("chained draw hidden when the callee's own gate (flag==0) fails")
-
-
-# ── a toggle can reassign the diffuse texture instead of (or as well as)
-#    gating a draw -- e.g. HousekeeperColumbina.ini's `$seven2` swap
-
-DIFFUSE_SWAP_INI = """[Constants]
-global persist $seven2 = 0
-
-[KeySeven2]
-key = k
-type = cycle
-$seven2 = 0,1
-
-[TextureOverrideColumbinaBodyBlend]
-ib = ResourceColumbinaBodyIB
-vb0 = ResourcePos
-vb1 = ResourceTc
-Resource\\GIMI\\Diffuse = ref ResourceDiffuseA
-drawindexed = 10, 0, 0
-if $seven2 == 1
-Resource\\GIMI\\Diffuse = ref ResourceDiffuseB
-else
-Resource\\GIMI\\Diffuse = ref ResourceDiffuseC
-endif
-drawindexed = 20, 100, 0
-
-[ResourceColumbinaBodyIB]
-filename = body.ib
-format = DXGI_FORMAT_R32_UINT
-
-[ResourcePos]
-filename = pos.buf
-stride = 40
-
-[ResourceTc]
-filename = tc.buf
-stride = 20
-
-[ResourceDiffuseA]
-filename = diffuseA.dds
-
-[ResourceDiffuseB]
-filename = diffuseB.dds
-
-[ResourceDiffuseC]
-filename = diffuseC.dds
-"""
-
-
-def _material_case_toggle_driven_diffuse_swap_ini_parser():
-    with tempfile.TemporaryDirectory() as tmp:
-        path = write(tmp, "mod.ini", DIFFUSE_SWAP_INI)
-        secs = merge_sections([path])
-        groups = build_draw_groups(secs, extract_resources(secs))
-        assert (len(groups) == 1), (f"one draw group built (got {len(groups)})")
-        group = groups[0]
-        assert (group["diffuse_file"] == "diffuseA.dds"), (f"group default diffuse is the first-seen, unconditional one "
-              f"(got {group['diffuse_file']})")
-        draws = {d["count"]: d for d in group["draws"]}
-
-        assert ("texture_variants" not in draws[10]), ("the earlier unconditional diffuse assignment doesn't leak a "
-              "spurious single-entry texture_variants onto the first draw")
-
-        variants = draws[20].get("texture_variants")
-        assert (bool(variants) and len(variants) == 2), (f"exactly 2 variants for the later, toggle-gated draw (got {variants})")
-        by_file = {v["file"]: v["conditions"] for v in variants}
-        assert (set(by_file) == {"diffuseB.dds", "diffuseC.dds"}), (f"both branches' files are present (got {sorted(by_file)})")
-        assert (_visible(by_file["diffuseB.dds"], {"seven2": "1"}) and
-              not _visible(by_file["diffuseB.dds"], {"seven2": "0"})), ("diffuseB's condition matches only seven2==1")
-        assert (_visible(by_file["diffuseC.dds"], {"seven2": "0"}) and
-              not _visible(by_file["diffuseC.dds"], {"seven2": "1"})), ("diffuseC's condition (the else branch) matches only seven2==0")
-
-
-def _material_case_toggle_driven_diffuse_swap_mesh_builder():
-    with tempfile.TemporaryDirectory() as tmp:
-        path = write(tmp, "mod.ini", DIFFUSE_SWAP_INI)
-        open(os.path.join(tmp, "body.ib"), "wb").write(
-            struct.pack("<3I", 0, 1, 2) + struct.pack("<3I", 3, 4, 5))
-        with open(os.path.join(tmp, "pos.buf"), "wb") as f:
-            for i in range(8):
-                f.write(struct.pack("<3f", float(i), float(i), float(i)) + b"\0" * 28)
-        open(os.path.join(tmp, "tc.buf"), "wb").write(b"\0" * 20 * 8)
-        for name in ("diffuseA.dds", "diffuseB.dds", "diffuseC.dds"):
-            open(os.path.join(tmp, name), "wb").write(b"DDS " + name.encode())
-
-        secs = merge_sections([path])
-        groups = build_draw_groups(secs, extract_resources(secs))
-        for d in groups[0]["draws"]:
-            d["start"], d["count"] = (0, 3) if d["count"] == 10 else (3, 3)
-        meshes, _geometry = build_mesh_fixture(groups, tmp)
-        by_draw = {tuple(e["drawindexed"]): e for e in meshes.values()}
-        first  = by_draw[(3, 0, 0)]
-        second = by_draw[(3, 3, 0)]
-
-        assert ("texture_variants" not in first), ("first draw's payload entry carries no texture_variants (single diffuse)")
-
-        variants = second.get("texture_variants")
-        assert (bool(variants) and len(variants) == 3 and
-              variants[0]["conditions"] == []), (f"second draw carries the unconditional write and both resolved "
-              f"conditional writes in source order (got {variants})")
-        keys = {texture_file(v["tex_key"]) for v in variants}
-        assert (keys == {"diffuseA.dds", "diffuseB.dds", "diffuseC.dds"}), (f"each assignment's tex_key names its own resolved diffuse file (got {keys})")
-        assert (texture_file(second["tex_key"]) == "diffuseB.dds"), (f"the draw's own default tex_key is the first/`if`-branch "
-              f"alternative at this point in execution order (seven2==1 -> "
-              f"diffuseB), not the group's earlier unconditional diffuseA "
-              f"(got {second['tex_key']})")
-
-
-# ── A single section reassigns the diffuse several times in execution order
-#    -- draws before the first assignment get
-#    none, draws between the first and second assignment get the first, draws
-#    after the second (unconditional) reassignment get that one, and draws
-#    after a THIRD (also unconditional) reassignment back to the first file
-#    get that again. A static one-diffuse-per-component model can only ever
-#    get 3 of these 4 regions right.
-
-SAME_VAR_PARTIAL_DIFFUSE_INI = """[Constants]
-global persist $color = 0
-
-[KeyColor]
-key = c
-type = cycle
-$color = 0,1,2
-
-[TextureOverrideBodyPosition]
-vb0 = ResourceBodyPosition
-
-[TextureOverrideBodyTexcoord]
-vb1 = ResourceBodyTexcoord
-
-[TextureOverrideBody]
-ib = ResourceBodyIB
-Resource\\GIMI\\Diffuse = ref ResourceDiffuseA
-if $color == 1
-Resource\\GIMI\\Diffuse = ref ResourceDiffuseB
-endif
-if $color == 2
-Resource\\GIMI\\Diffuse = ref ResourceDiffuseC
-endif
-drawindexed = 3, 0, 0
-
-[ResourceBodyIB]
-filename = body.ib
-[ResourceBodyPosition]
-filename = pos.buf
-stride = 40
-[ResourceBodyTexcoord]
-filename = tc.buf
-stride = 20
-[ResourceDiffuseA]
-filename = diffuseA.dds
-[ResourceDiffuseB]
-filename = diffuseB.dds
-[ResourceDiffuseC]
-filename = diffuseC.dds
-"""
-
-
-def _material_case_same_variable_partial_diffuse_chains_keep_assignment_history():
-    with tempfile.TemporaryDirectory() as tmp:
-        path = write(tmp, "mod.ini", SAME_VAR_PARTIAL_DIFFUSE_INI)
-        secs = merge_sections([path])
-        draw = build_draw_groups(secs, extract_resources(secs))[0]["draws"][0]
-        assignments = draw.get("texture_assignments") or []
-        assert ([item["file"] for item in assignments] ==
-              ["diffuseA.dds", "diffuseB.dds", "diffuseC.dds"]), (f"independent same-variable writes retain source order (got {assignments})")
-
-        def selected(value):
-            state = {"color": value}
-            return next((item["file"] for item in reversed(assignments)
-                         if _visible(item["conditions"], state)), None)
-
-        assert ([selected(value) for value in ("0", "1", "2")] ==
-              ["diffuseA.dds", "diffuseB.dds", "diffuseC.dds"]), ("last-matching assignment selects the authored texture for every color")
-
-
-MULTI_REASSIGN_INI = """[KeySuitCL]
-key = l
-type = cycle
-$SuitCL = 0,1
-
-[TextureOverrideMultiPosition]
-vb0 = ResourceMultiPosition
-
-[TextureOverrideMultiBlend]
-vb1 = ResourceMultiBlend
-
-[TextureOverrideMultiTexcoord]
-vb1 = ResourceMultiTexcoord
-
-[TextureOverrideMultiA]
-ib = ResourceMultiIB
-drawindexed = 10, 0, 0
-if $SuitCL == 0
-Resource\\ZZMI\\Diffuse = ref ResourceDiffuseA
-elif $SuitCL == 1
-Resource\\ZZMI\\Diffuse = ref ResourceDiffuseA2
-endif
-drawindexed = 20, 10, 0
-Resource\\ZZMI\\Diffuse = ref ResourceDiffuseB
-drawindexed = 30, 30, 0
-Resource\\ZZMI\\Diffuse = ref ResourceDiffuseA
-drawindexed = 40, 60, 0
-
-[ResourceMultiIB]
-filename = multi.ib
-format = DXGI_FORMAT_R32_UINT
-
-[ResourceMultiPosition]
-filename = pos.buf
-stride = 40
-
-[ResourceMultiBlend]
-filename = blend.buf
-stride = 32
-
-[ResourceMultiTexcoord]
-filename = tc.buf
-stride = 20
-
-[ResourceDiffuseA]
-filename = diffuseA.dds
-
-[ResourceDiffuseA2]
-filename = diffuseA2.dds
-
-[ResourceDiffuseB]
-filename = diffuseB.dds
-"""
-
-
-def _material_case_multi_reassignment_diffuse_resolution():
-    with tempfile.TemporaryDirectory() as tmp:
-        path = write(tmp, "mod.ini", MULTI_REASSIGN_INI)
-        secs = merge_sections([path])
-        groups = build_draw_groups(secs, extract_resources(secs))
-        assert (len(groups) == 1), (f"one draw group built (got {len(groups)})")
-        group = groups[0]
-        by_start = {d["start"]: d for d in group["draws"]}
-
-        assert (by_start[0].get("texture_default_file") is None), (f"draw before any diffuse assignment gets none "
-              f"(got {by_start[0].get('texture_default_file')})")
-        assert (by_start[10].get("texture_default_file") == "diffuseA.dds"), (f"draw after the if/elif chain's first branch gets diffuseA "
-              f"(got {by_start[10].get('texture_default_file')})")
-        variants = by_start[10].get("texture_variants")
-        assert (bool(variants) and len(variants) == 2), (f"that same draw also carries both toggle alternatives "
-              f"(got {variants})")
-        assert (by_start[30].get("texture_default_file") == "diffuseB.dds"), (f"draw after the unconditional reassignment to B gets B, not "
-              f"the earlier if/elif chain's A "
-              f"(got {by_start[30].get('texture_default_file')})")
-        assert ("texture_variants" not in by_start[30]), ("the unconditional B reassignment carries no toggle variants")
-        assert (by_start[60].get("texture_default_file") == "diffuseA.dds"), (f"draw after reassigning back to A gets A again, not stuck on "
-              f"B from the earlier unconditional reassignment "
-              f"(got {by_start[60].get('texture_default_file')})")
-
-        pool = [p["res"] for p in group["diffuse_pool_files"]]
-        assert (pool == ["ResourceDiffuseA", "ResourceDiffuseA2", "ResourceDiffuseB"]), (f"the group's texture pool lists every distinct diffuse "
-              f"referenced anywhere in the section, in first-seen order "
-              f"(got {pool})")
-
-
-def _material_case_multi_reassignment_mesh_builder():
-    with tempfile.TemporaryDirectory() as tmp:
-        path = write(tmp, "mod.ini", MULTI_REASSIGN_INI)
-        open(os.path.join(tmp, "multi.ib"), "wb").write(
-            struct.pack("<12I", *range(12)))
-        with open(os.path.join(tmp, "pos.buf"), "wb") as f:
-            for i in range(12):
-                f.write(struct.pack("<3f", float(i), float(i), float(i)) + b"\0" * 28)
-        open(os.path.join(tmp, "tc.buf"), "wb").write(b"\0" * 20 * 12)
-        for name in ("diffuseA.dds", "diffuseA2.dds", "diffuseB.dds"):
-            open(os.path.join(tmp, name), "wb").write(b"DDS " + name.encode())
-
-        secs = merge_sections([path])
-        groups = build_draw_groups(secs, extract_resources(secs))
-        # Remap onto small, non-overlapping windows of the 12-index buffer --
-        # the ini's own start/count values (0/10/30/60) are execution-order
-        # markers only, not real offsets into this tiny fixture buffer.
-        orig_starts = [d["start"] for d in groups[0]["draws"]]
-        for i, d in enumerate(groups[0]["draws"]):
-            d["start"], d["count"] = i * 3, 3
-        meshes, _geometry = build_mesh_fixture(groups, tmp)
-        by_orig_start = dict(zip(orig_starts,
-            sorted(meshes.values(), key=lambda e: e["drawindexed"][1])))
-
-        assert (by_orig_start[0]["tex_key"] is None), (f"first draw's resolved tex_key is None (got {by_orig_start[0]['tex_key']})")
-        assert (texture_file(by_orig_start[10]["tex_key"]) == "diffuseA.dds"), (f"second draw resolves diffuseA (got {by_orig_start[10]['tex_key']})")
-        assert (texture_file(by_orig_start[30]["tex_key"]) == "diffuseB.dds"), (f"third draw resolves diffuseB, not the earlier diffuseA "
-              f"(got {by_orig_start[30]['tex_key']})")
-        assert (texture_file(by_orig_start[60]["tex_key"]) == "diffuseA.dds"), (f"fourth draw resolves back to diffuseA "
-              f"(got {by_orig_start[60]['tex_key']})")
-
-        options = by_orig_start[0].get("texture_options")
-        assert (bool(options) and len(options) == 3), (f"every draw in the component carries the full 3-entry "
-              f"texture pool for the manual picker (got {options})")
-
-
-# ── A section with NO drawindexed line at all (the implicit whole-buffer
-#    draw -- see HANDLING_SKIP_INI above) still assigns a diffuse via a bare `ps-t0 =` slot.
-#    build_draw_groups' synthetic placeholder draw must carry that forward,
-#    not silently drop it -- it used to hardcode an empty diff_variants list
-#    for this placeholder, leaving the whole component untextured even
-#    though the section plainly names a diffuse.
-
-IMPLICIT_DRAW_DIFFUSE_INI = """[TextureOverrideImplicitPosition]
-vb0 = ResourceImplicitPosition
-
-[TextureOverrideImplicitBlend]
-vb1 = ResourceImplicitBlend
-
-[TextureOverrideImplicitTexcoord]
-vb1 = ResourceImplicitTexcoord
-
-[TextureOverrideImplicitA]
-ib = ResourceImplicitIB
-ps-t0 = ResourceImplicitDiffuse
-
-[ResourceImplicitIB]
-filename = implicit.ib
-format = DXGI_FORMAT_R32_UINT
-
-[ResourceImplicitPosition]
-filename = pos.buf
-stride = 40
-
-[ResourceImplicitBlend]
-filename = blend.buf
-stride = 32
-
-[ResourceImplicitTexcoord]
-filename = tc.buf
-stride = 20
-
-[ResourceImplicitDiffuse]
-filename = implicit.dds
-"""
-
-
-def _mesh_case_implicit_whole_buffer_draw_keeps_its_diffuse():
-    with tempfile.TemporaryDirectory() as tmp:
-        path = write(tmp, "mod.ini", IMPLICIT_DRAW_DIFFUSE_INI)
-        secs = merge_sections([path])
-        groups = build_draw_groups(secs, extract_resources(secs))
-        assert (len(groups) == 1), (f"one draw group built (got {len(groups)})")
-        group = groups[0]
-        assert (len(group["draws"]) == 1), ("exactly the one synthetic placeholder draw")
-        draw = group["draws"][0]
-        assert (draw.get("count") is None), ("the placeholder draw has no count -- it's the implicit whole-buffer read")
-        assert (draw.get("texture_default_file") == "implicit.dds"), (f"the placeholder draw still resolves the section's own "
-              f"ps-t0 diffuse, not None (got {draw.get('texture_default_file')})")
-
-
-# ── XXMI-generated mods assign the diffuse without the "ref" keyword
-#    (e.g. "Resource\GIMI\Diffuse = ResourceXDiffuse"), unlike the
-#    "= ref X" form other tools emit -- both must resolve the same way.
-
-DIFFUSE_NO_REF_INI = """[TextureOverrideXPosition]
-vb0 = ResourceXPosition
-
-[TextureOverrideXBlend]
-vb1 = ResourceXBlend
-
-[TextureOverrideXTexcoord]
-vb1 = ResourceXTexcoord
-
-[TextureOverrideXA]
-ib = ResourceXAIB
-Resource\\GIMI\\Diffuse = ResourceXDiffuse
-run = CommandList\\GIMI\\SetTextures
-drawindexed = 100, 0, 0
-
-[ResourceXPosition]
-filename = pos.buf
-stride = 40
-
-[ResourceXBlend]
-filename = blend.buf
-stride = 32
-
-[ResourceXTexcoord]
-filename = tc.buf
-stride = 20
-
-[ResourceXAIB]
-filename = a.ib
-format = DXGI_FORMAT_R32_UINT
-
-[ResourceXDiffuse]
-filename = diffuseX.dds
-"""
-
-
-def _material_case_diffuse_assignment_without_ref_keyword():
-    with tempfile.TemporaryDirectory() as tmp:
-        path = write(tmp, "mod.ini", DIFFUSE_NO_REF_INI)
-        secs = merge_sections([path])
-        groups = build_draw_groups(secs, extract_resources(secs))
-        names = {g["display_name"]: g for g in groups}
-        assert ("XA" in names), (f"the draw section builds a group (got {sorted(names)})")
-        if "XA" in names:
-            assert (names["XA"]["diffuse_file"] == "diffuseX.dds"), (f"the no-\"ref\" diffuse assignment still resolves "
-                  f"(got {names['XA']['diffuse_file']})")
-
-
-# ── index buffers come in both DXGI_FORMAT_R32_UINT and _R16_UINT; reading a
-#    16-bit one as 32-bit yields garbage indices and no usable mesh.
-
-IB_R16_INI = DIFFUSE_NO_REF_INI.replace("DXGI_FORMAT_R32_UINT", "DXGI_FORMAT_R16_UINT")
-
-
-def _mesh_case_r16_index_buffer():
-    with tempfile.TemporaryDirectory() as tmp:
-        path = write(tmp, "mod.ini", IB_R16_INI)
-        open(os.path.join(tmp, "a.ib"), "wb").write(struct.pack("<3H", 5, 6, 7))
-        with open(os.path.join(tmp, "pos.buf"), "wb") as f:
-            for i in range(8):
-                f.write(struct.pack("<3f", float(i), float(i), float(i)) + b"\0" * 28)
-        open(os.path.join(tmp, "tc.buf"), "wb").write(b"\0" * 20 * 8)
-
-        secs = merge_sections([path])
-        groups = build_draw_groups(secs, extract_resources(secs))
-        assert (groups and groups[0]["index_size"] == 2), (f"an R16_UINT ib reports 2 bytes per index "
-              f"(got {groups[0]['index_size'] if groups else None})")
-        for d in groups[0]["draws"]:
-            d["count"] = 3
-        meshes, geometry = build_mesh_fixture(groups, tmp)
-        assert (len(meshes) == 1), (f"the draw builds a mesh (got {len(meshes)})")
-        entry = next(iter(meshes.values()))
-        pos = geometry_values(geometry, entry["pos"])
-        verts = sorted(round(pos[i]) for i in range(0, len(pos), 3))
-        assert (verts == [5, 6, 7]), (f"16-bit indices are decoded as 16-bit, not 32-bit (got {verts})")
